Deletion for BTrees.Cursor also sets out-variables for key, value, position.
The reason is mainly that this eases the BLOB BTrees.
However, in other scenarios this might make sense, too, if one has called Next
and then hands over the Cursor to some more general procedure that will call a
deletion and wants to know about key, value and/or position.