Several rather minor changes.
DB.Types.Values is not just an instance of a Gen_Numbers anymore but
a full package.
The content is still the same, but the type is now Value_Type instead
of Number_Type (this is the main reason for this step).
Furthermore, From/To_Storage_Array are implemented.

DB.Gen_Heaps formal generic parameter Item_Type may be constrained at
the moment. Maybe this should be undone.

DB.BTrees and DB.Blob_Trees Cursor_Type's Bound_Type is unconstrained
in the public part of the specs now. It's still constrained in the
private part, but has a default parameter. This allows us to used in
other records.