New coding style: `is<LF>begin' instead of `is begin'.
Hence it is
   function X
     (Y : Y_Type)
      return X_Type is
   begin
     ...
or
   function X
     (Y : Y_Type)
      return X_Type
   is
      Z : Z_Type;
   begin
     ...
Analoguously for procedures and entries.

In the second case, the `is' has its own line to draw a line between the spec
and the body.

This seems to be very common style. In contrast to this style, we put the
return type of functions in a separate line instead of the same line as the
last parameter.