Added Key/Value_Size_Bound to Gen_BTrees.
The Key/Value_Size_Bound results are not used calculations. Their return values
are however checked in the Insert procedures so that never a too large key/value
pair can be inserted into a tree, because such a key/value pair can lead to
violations of the tree constraints much later.

Furthermore, the Gen_BTrees package and its users are Preelaborated which has
the effect that in the nested package Nodes the constants had to be redefined
as functions since their values are not static.