Fixed: Swap_Nodes failed due to duplicate-key entries.
The reason was pretty simple: Key_Position was used in the Prepare_References
procedure when the child-pointer of the parent is adjusted. Now there
is not only one address parameter (the new address of the node) but also the
old one. For sake of understanding, the parameters have Old_N_A and New_N_A
as names.