Fixed wrong boundaries of sizes of test data. To be checked twice!
In the old code, the Max_String_Length and Max_Key_Size values in the
test drivers were probably wrong. First of all, the Size_Of_Position was
not taken into account which is needed per-entry as internal node overhead.
Second, the value (4k - Size_Of_Meta_Data) / 4 was wrong (1012 or so, now
1020), because Size_Of_Meta_Data = 16.

Just check them again several times and try with +/- 1 values.