Removed symb. traceback. Replaced Height_Type by Natural. Plan for prefix-compr.
Symbolic tracebacks can be got using addr2line for the respective binary.

Now the libdb also compiles as relocatable library, but then you don't have
symbolic traceback. It seems that using "ldd" to get the base address of
libdb and then subtracting this base address from the exception's occurence's
address, you get the address of the location in the shared library. But this
didn't work for me yet :(.

See TODO for plans for prefix compression. It involves new parameters for
Gen_BTrees & Co, which allow the Write_* procedures to construct a binary
tree in thereby reduce the number of decodings needed.