First untested version of an efficient Gen_Mutex_Sets.
Is it really thread-safe?

Furthermore found a bug in the B-tree: in Move_Right we don't know
how long to move right due to possible duplicates of keys.
I think this could be fixed if we not only look for the key but also
for the child pointer (in the inner-node case) we're looking for. Maybe
the child pointer alone is already enough, isn't it?
For the leaf-nodes we'll need a separate Move_Right that doesn't care
about this stuff.