Fixed concurrency errors.
The Pop_Leaf and Pop_Inner procedures fix the stack when they tried
to access to the root node and the root node has changed (i.e. the
tree has grown) since the stack was initially built.

Postgres does the same according to their README in the nbtree or
so folder.

Note: there still seems to be a deadlock.
In one run of ./rundouble.sh only 28 of the second insertion sequence
terminated, then we probably had an infinite loop.