Fixed: concurrency issues.
The reason was that R_A was unlocked in Insert_Key_And_Update_High_Key
too early: if a Rebuild_Stack was in progress at the same time which had
C_A = R_A and the parent pointing to R_A wasn't written yet, the
Rebuild_Stack never found the node that has a pointer to C_A = R_A.

The Unlock(R_A) could be done earlier in the ascend.

Maybe there's also another way to fix this at another point:
in Write_And_Ascend's splitting branch. Does it help to do the ascend
first and then write L back? We wouldn't need to lock R_A then at all,
but then we would have to hold the lock to L_A until the ascend has
completed, so this is probably not any better either.