Alternative fix to concurrency issues.
From commit 1a83b61fb477598775490dcc7daf3cd87db525da (fix #1):
The reason was that R_A was unlocked in Insert_Key_And_Update_High_Key
too early: if a Rebuild_Stack was in progress at the same time which had
C_A = R_A and the parent pointing to R_A wasn't written yet, the
Rebuild_Stack never found the node that has a pointer to C_A = R_A.

Now the Unlock(Tree, R_A) is performed directly as it was earlier,
but the (Re-) Build_Stack procedure has changed (see documentation).

The problem with the insertion of C_A to a parent and the concurrent
search for a node containing the pointer to C_A should be solved because
of the use of Move_Right which is thread-safe and therefore waits for
C_A being inserted.

Also did some refactoring; most notably introduced nested Stack package.