Build_Stack: Root_Address read twice, lock only in second.
Before, the root was locked, then read, then unlocked, because we want
to ensure that it really is the root, which is the case if the link is
invalid.
Now we read it without lock. In most cases, the link is invalid and
we're fine. If the link is valid, we lock the root, re-read it and
unlock it.