Insertion: insert into next leaf entry doesn't fit
If the high key of the current leaf is the same as to-be-inserted key,
we can also insert the entry in the next leaf.

We exploit this to safe space (when duplicate keys are allowed):
If the entry doesn't fit into the current node, we try to insert it in
the next leaf etc.