Fixed: node-internal-order errors due to high keys
The fix from commit 5580ad9686fc6dff2cadbe3b0f69bef3b5b3900c lead to
invalid-link-orders for some reason, I haven't understood it yet or
forgot it.

The new attempt to fix it is in the delete procedure: we adjust the the
keys of the parents if the high keys of the nodes change.