Block_IO.Write has Cache_Priority parameter.
In the Gen_BTree, the Cache_Priority is set to the level of the node.
Since leaves have the lowest level, the higher nodes get higher
priority.

The idea is not to use the Cache_Priority yet in the Local_IO, because
there the OS does the caching and for some reason, it does a pretty good
job.
But for the Network_IO, we'll have to write our own cache and this
should exploit the Cache_Priority.