Some small code improvements.
* Some more Preelaboate pragmas
* Some more Pure pragmas
* New_Map: * additional overloading with 'implementation' parameter
           * added Allow_Duplicates parameter

Tried shifting DB.Gen_BTrees to DB.Maps.Gen_BTrees.
This works well, but then either DB.Gen_Map_Reduce needs to access
DB.Maps.Gen_BTrees which is not nice because DB.Maps.Gen_BTrees should
be a private unit, or it has to use the DB.Maps interface which is
probably bad for many applications because they don't need the bloat of
the default Key_Type.