Fixed: BTree Cursor: broken Search_Lower_Bound and Recalibrate
The Search_Node procedure's outcome was only useful if it found an exact
match. This is the right behaviour for the normal Search operation, but
not for Cursor's Search_Lower_Bound and Recalibrate.
For these we need to search the minimum key that is greater or equal
than the parameter key. Then we move to the right to satisfy the
condition.
Care needs to be taken to handle empty nodes correctly. The Search_Node
now still sets State to Success only if an exact match is found, but it
sets Index in all cases. If Index <= degree of the node, it has found a
greater (State = Failure) or equal (State = Success) key. Otherwise,
the fact that Index > degree of the node indicates that the minimum
greater key is in a node right from the current (or doesn't exist).