Fixed: bug in regular expression operation Union
The union of two DFAs has a new alphabet, i.e. a new mapping.
All DFAs have a default-column, 0, which is used for all characters
in a test-string that are not in the DFA's mapping.
When we create the union of two DFAs, we also consider these
default-columns, of course.
But we need to take care: when there are characters now in the
mapping that weren't before, these won't fall into the default-column
class at runtime of the matching-test.
So we need to add additional transitions for these characters which
are now in the mapping but weren't before (to the same state to which
the DFA's default-columns direct to in the union, of course).