Fixed package structure (see last commit); use of private-with.
The integer and string Value_Type implementations should not have been
children of DB.Maps.Values, because DB.Maps.Values contains the
serialization of Value_Type objects which has nothing to do with the
Value_Type interface itself, while the sample implementations are in
fact implementations of the Value_Type.

Due to the use of "private with", the code is less Ada 95 compliant.