Maps: added Value_Wrapper_Type operations which have out Value_Type.
Before, the user had to provide a prototype of the right Value_Type
implementation for operations that have a out-parameter of type
Value_Type, i.e. Map.Search, Map.Search_Minimum, Map.Delete,
Cursor.Next and Cursor.Delete.

The Value_Wrapper_Type was before used in the Value_IO_Signature of
the Gen_BTrees instance in Maps.Bounded.
Hence, it lived in the Maps.Abstract_Value_Serialization and was used
only in Maps.Bounded.
Now, it is moved to Maps and is intended to be used from outside.
It is implemented generically in DB.Utils.Gen_Auto_Pointers.

I hope I haven't introduced any memory leaks by this.