Gen_BTrees: API has now Insert, Replace, Append.
The procedures Insert, Replace, Append replace the old Insert in
combination with Allow_Duplicates.
The semantics is as follows:
 * Insert fails if the Key already exists.
   This is putIfAbsent() in Java.
 * Append adds the new Key/Value pair, even if the Key already
   exists. The old Key/value pairs are all kept.
 * Replace inserts or, if Key already exists, replaces the Value
   associated with Key.
   This is put() in Java.
Furthermore there are versions of Insert and Replace which have
parameters Existed and Old_Value to indicate the existing value.

Adjusted the tests.

One should still write some unit tests that quickly test the variants
for stupid copy-paste-or-so-errors.