Gen_BTrees: value is written with restricted cursor
The cursor is restricted to the maximum allowed value size (with respect
to the key).
Hence, if a bigger value is written, an error is raised. Also, using
Blocks.Remaining_Space, the maximum remaining space for the value can be
queried.

This is intended to allow values to be written partially with the rest
in the heap file.
This mechanism would hence live in the serialization of values!